+++ public/app.js
@@ -1,11 +1,11 @@
 import makePlayer from './players.js'
-import makeClassement from './classement.js'
 import navbar from './navbar.js'
 import calendrier from './calendrier.js';
 import formulaire from './formulaire.js'
 import admin from './admin.js'
 import newRace from './newrace.js'
 import information from './information.js'
+import classementRoute  from './classementroute.js';
 const mainDiv = document.getElementById('main')
  
 const render = html => {
@@ -39,50 +39,7 @@ const controllers = {
     ),
     '/calendrier': calendrier
     ,
-    '/classement': () =>
-      fetch('/courses')
-      .then(res => res.json())
-      .then(races => races.reduce((carry, race) => carry + makeClassement(race),''))
-      .then(gpCard => {
-        render(
-        `${navbar}
-        <div class="container">
-          <div class="jumbotron ImageClassement">
-            <h1 class="display-3 titleClassement">Classement</h1>
-            <p></p>
-            </div>
-            <div class="row">${gpCard}</div>
-        </div>
-        `)
-        // document.get...
-        const buttonPositions = document.getElementsByClassName('add-position')
-        for (let buttonPosition of buttonPositions) {
-          //console.log(buttonPosition)
-          buttonPosition.addEventListener('click', () => {
-            const raceId = buttonPosition.dataset.raceId
-            const playerId = buttonPosition.dataset.playerId
-            const position = buttonPosition.dataset.position
-            console.log(raceId, playerId, position)
-            const addPositionToData = {
-              race_id: raceId,
-              player_id: playerId,
-              position: position
-            }
-            fetch('/courses', {
-              method: 'PUT',
-              headers: {
-                'Accept': 'application/json, text/plain, */*',
-                'Content-Type': 'application/json'
-              },
-              body: JSON.stringify(addPositionToData)
-            },
-           // window.location = "/calendrier"
-            )
-            
-          })
-        }
-      }
-    ),
+    '/classement': classementRoute,
 
     '/race/new': newRace,
